+++ python/dgl/graphbolt/__init__.py
@@ -5,7 +5,6 @@ import sys
 import torch
 
 from .._ffi import libinfo
-from .graph_storage import *
 from .itemset import *
 from .minibatch_sampler import *
 from .feature_store import *
@@ -15,7 +14,8 @@ from .dataset import *
 from .impl import *
 from .dataloader import *
 from .subgraph_sampler import *
-from .sampled_sub_graph import *
+from .sampled_subgraph import *
+
 
 def load_graphbolt():
     """Load Graphbolt C++ library"""
@@ -45,4 +45,4 @@ def load_graphbolt():
 
 load_graphbolt()
 
-SampledSubgraph = torch.classes.graphbolt.SampledSubgraph
+C_SampledSubgraph = torch.classes.graphbolt.SampledSubgraph
