--- python/dgl/graphbolt/graph_storage/__init__.py
@@ -1,2 +1,0 @@
-"""Graphbolt graph module."""
-from .csc_sampling_graph import *
