+++ python/dgl/graphbolt/impl/__init__.py
@@ -2,3 +2,5 @@
 from .ondisk_dataset import *
 from .ondisk_metadata import *
 from .torch_based_feature_store import *
+from .csc_sampling_graph import *
+from .csc_sampling_graph_sampled_subgraph import *
